requirements.txt
streamlit
pandas
plotly
fpdf2
passlib